refactor: reorganize project structure and standardize test suites
- move heapq module to lib/ and update makefile paths
- improve makefile test runner to properly count failed tests
- replace legacy test prints with test_group and new assert macros
- encapsulate test logic into static functions across all modules
- standardize assertion messages to use the "handle" convention
- add null pointer safeguard

+++ coders/include/coder.h
@@ -6,7 +6,7 @@
 /*   By: nlallema <nlallema@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/03/25 11:49:16 by nlallema          #+#    #+#             */
-/*   Updated: 2026/03/31 12:07:36 by nlallema         ###   ########lyon.fr   */
+/*   Updated: 2026/03/31 15:01:07 by nlallema         ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -39,7 +39,7 @@ typedef struct s_coder
 typedef struct s_coder_array
 {
 	t_coder			*items;
-	int				count;
+	size_t			count;
 	pthread_mutex_t	start_mutex;
 	int				start_mutex_init;
 	pthread_cond_t	start_cond;
